+++ android/src/main/java/com/example/android_flutter_wifi/wifiUtils/Logger.java
@@ -1,0 +1,5 @@
+package com.example.android_flutter_wifi.wifiUtils;
+
+public interface Logger {
+    void log(int priority, String tag, String message);
+}