+++ android/src/main/java/com/example/android_flutter_wifi/wifiUtils/TypeEnum.java
@@ -1,0 +1,13 @@
+package com.example.android_flutter_wifi.wifiUtils;
+
+/**
+ * Author:  Minuit
+ * Time: 2020/12/1
+ * Description:
+ */
+public enum TypeEnum {
+    OPEN,
+    WEP,
+    PSK,
+    EAP
+}