+++ android/src/main/java/com/example/android_flutter_wifi/wifiUtils/wifiDisconnect/DisconnectionErrorCode.java
@@ -1,0 +1,7 @@
+package com.example.android_flutter_wifi.wifiUtils.wifiDisconnect;
+
+public enum DisconnectionErrorCode {
+    COULD_NOT_GET_WIFI_MANAGER,
+    COULD_NOT_GET_CONNECTIVITY_MANAGER,
+    COULD_NOT_DISCONNECT,
+}