+++ android/src/main/java/com/example/android_flutter_wifi/wifiUtils/wifiDisconnect/DisconnectionSuccessListener.java
@@ -1,0 +1,9 @@
+package com.example.android_flutter_wifi.wifiUtils.wifiDisconnect;
+
+import androidx.annotation.NonNull;
+
+public interface DisconnectionSuccessListener {
+    void success();
+
+    void failed(@NonNull DisconnectionErrorCode errorCode);
+}