+++ android/src/main/java/com/example/android_flutter_wifi/wifiUtils/wifiScan/WifiScanCallback.java
@@ -1,0 +1,5 @@
+package com.example.android_flutter_wifi.wifiUtils.wifiScan;
+
+public interface WifiScanCallback {
+    void onScanResultsReady();
+}