+++ android/src/main/java/com/example/android_flutter_wifi/wifiUtils/wifiState/WifiStateCallback.java
@@ -1,0 +1,5 @@
+package com.example.android_flutter_wifi.wifiUtils.wifiState;
+
+public interface WifiStateCallback {
+    void onWifiEnabled();
+}