+++ android/src/main/java/com/example/android_flutter_wifi/wifiUtils/wifiWps/ConnectionWpsListener.java
@@ -1,0 +1,5 @@
+package com.example.android_flutter_wifi.wifiUtils.wifiWps;
+
+public interface ConnectionWpsListener {
+    void isSuccessful(boolean isSuccess);
+}