+++ example/lib/main.dart
@@ -26,6 +26,9 @@ class _MyAppState extends State<MyApp> {
   Widget build(BuildContext context) {
     return MaterialApp(
       home: Scaffold(
+        floatingActionButton: FloatingActionButton(onPressed: () {
+          connectToNetworkWithSSID();
+        }),
         appBar: AppBar(
           title: const Text('Plugin example app'),
         ),
@@ -39,10 +42,14 @@ class _MyAppState extends State<MyApp> {
   void init() async {
     await AndroidFlutterWifi.init();
     var isConnected = await AndroidFlutterWifi.isConnected();
-    print('Is connected: ${isConnected.toString()}');
+    await getDhcpInfo();
+    await getWifiList();
+    await getConfiguredNetworks();
+    //await forgetWifi();
+    debugPrint('Is connected: ${isConnected.toString()}');
   }
 
-  void getWifiList() async {
+  getWifiList() async {
     List<WifiNetwork> wifiList = await AndroidFlutterWifi.getWifiScanResult();
     if (wifiList.isNotEmpty) {
       WifiNetwork wifiNetwork = wifiList[0];
@@ -61,5 +68,52 @@ class _MyAppState extends State<MyApp> {
   getActiveWifiNetwork() async {
     ActiveWifiNetwork activeWifiNetwork =
         await AndroidFlutterWifi.getActiveWifiInfo();
+  }
+
+  getDhcpInfo() async {
+    DhcpInfo dhcpInfo = await AndroidFlutterWifi.getDhcpInfo();
+    String ipString = AndroidFlutterWifi.toIp(dhcpInfo.gateway!);
+    String formedIp = AndroidFlutterWifi.getFormedIp(ipString);
+    print('Gateway: ${ipString}');
+    print('Formed ip: ${formedIp}');
+  }
+
+  void connectionTest() async {
+    String ssid = 'TP-Link_F9D0';
+    String password = ''; //Add you password
+    if (ssid.isEmpty) {
+      throw ("SSID can't be empty");
+    }
+    if (password.isEmpty) {
+      throw ("Password can't be empty");
+    }
+    debugPrint('Ssid: $ssid, Password: $password');
+    var result = await AndroidFlutterWifi.connectToNetwork(ssid, password);
+
+    debugPrint('---------Connection result-----------: ${result.toString()}');
+  }
+
+  ///Returns list of saved wifi networks on your device
+  getConfiguredNetworks() async {
+    List<ConfiguredNetwork> list =
+        await AndroidFlutterWifi.getConfiguredNetworks();
+    for (var element in list) {
+      debugPrint('Network id: ${element.networkId}');
+    }
+  }
+
+  ///Forget current connection
+  forgetWifi() async {
+    ActiveWifiNetwork activeWifiNetwork =
+        await AndroidFlutterWifi.getActiveWifiInfo();
+    var result =
+        await AndroidFlutterWifi.forgetWifiWithSSID(activeWifiNetwork.ssid!);
+    debugPrint('Forget wifi result: ${result.toString()}');
+  }
+
+  void connectToNetworkWithSSID() async {
+    var result =
+        await AndroidFlutterWifi.connectToNetworkWithSSID('Xartic_85AC9C_5G');
+    debugPrint('Connection result: ${result.toString()}');
   }
 }